fix max_end_stability and p3core.pl

+++ src/libprimer3.c
@@ -6254,7 +6254,7 @@ p3_set_gs_primer_internal_oligo_mishyb_library(p3_global_settings * p , char * p
 }
 
 void
-p3_set_gs_primer_max_end_stability(p3_global_settings * p , int val) {
+p3_set_gs_primer_max_end_stability(p3_global_settings * p , double val) {
   p->max_end_stability = val ;
 }
